refactor(web): PR-3 Task 2 — useGameStore → useGameSessionStore 치환
gameStore import 을 모두 gameSessionStore 로 교체 (alias 유지).
selector 는 gameSelectors 배럴에서 가져오도록 정리. 컴포넌트 테스트
mock 도 gameSessionStore + gameSelectors 분리 mock 으로 갱신 (B-4).

호출자 11건:
- hooks/useGameSync.ts
- pages/GamePage.tsx
- features/game/components/{Accusation,Ending,GameChat,GameHUD,PhaseTransition,Reading,Vote,Voting}Panel.tsx
- __tests__/{Ending,GroupChat,Reading,Vote}Panel.test.tsx
